@@ -1,32 +1,29 @@
-import file from '../assets/icons/file.svg';
-import call from '../assets/icons/call.svg';
-import brief from '../assets/icons/brief.svg';
-import circle from '../assets/icons/circle.svg';
-import info from '../assets/icons/info.svg';
-import hamburger from '../assets/icons/hamburger.svg';
-import About from '../View/About';
-import Application from '../View/Application';
-import Dashboard from '../View/Dashboard';
-import Jobs from '../View/Jobs';
-import Menu from '../View/Menu';
-import Policy from '../View/Policy';
-import Quote from '../View/Quote';
-import Trucks from '../View/Trucks';
-import WithUs from '../View/WithUs';
-
-
-export const nav=[
-    { id: 1, title: 'Driver Application', src: file, path: '/driverapplication', Component: Application },
-    { id: 2, title: 'Call', src: call, path: '/call', },
-    { id: 3, title: 'Get Quote', src: brief, path: '/getquote', Component: Quote},
-    { id: 4, title: 'About us', src: info, path: '/aboutus', Component: About},
-    {id:5, src:hamburger, path:'/menu', Component: Menu},
-];
-export default nav;
-export const circlemenu = [
-    { id: 6, src: circle, path: '/', Component: Dashboard },
-    { id: 7, src: circle, path: '/jobs', Component: Jobs },
-    { id: 8, src: circle, path: '/trucks', Component: Trucks },
-    { id: 9, src: circle, path: '/withus', Component: WithUs },
-    { id: 10, src: circle, path: '/companypolicy', Component: Policy }
+export const data = [
+  { id: 1, name: "Tedda	Beecraft", email: "tbeecraft0@ihg.com" },
+  { id: 2, name: "Dorothea	Hufton", email: "dhufton1@over-blog.com" },
+  { id: 3, name: "Ringo	Cholton", email: "rcholton2@goo.ne.jp" },
+  { id: 4, name: "Dodi	Grzelczak", email: "dgrzelczak3@abc.net.au" },
+  { id: 5, name: "Caril	Mantram", email: "cmantram4@jalbum.net" },
+  { id: 6, name: "Jere	Yuill", email: "jyuill5@bandcamp.com" },
+  { id: 7, name: "Ginnie	Smoughton", email: "gsmoughton6@sun.com" },
+  { id: 8, name: "Ricardo	Sutherley", email: "rsutherley7@shutterfly.com" },
+  { id: 9, name: "Bobbi	Syrie", email: "bsyrie8@yahoo.com" },
+  { id: 10, name: "Emmott	Mueller", email: "emueller9@businessinsider.com" },
+  { id: 11, name: "Filia	Korneev", email: "fkorneeva@theatlantic.com" },
+  { id: 12, name: "Vonni	Yerill", email: "vyerillb@eventbrite.com" },
+  { id: 13, name: "Joelly	Le Cornu", email: "jlec@mozilla.com" },
+  { id: 14, name: "Margaret	Linny", email: "mlinnyd@chicagotribune.com" },
+  { id: 15, name: "Guillaume	Pirdy", email: "gpirdye@flickr.com" },
+  { id: 16, name: "Theodora	Child", email: "tchildf@utexas.edu" },
+  { id: 17, name: "Louis	Sodeau", email: "lsodeaug@alexa.com" },
+  { id: 18, name: "Wendi	Doniso", email: "wdonisoh@statcounter.com" },
+  { id: 19, name: "Claudie	Levensky", email: "clevenskyi@hibu.com" },
+  { id: 20, name: "Kala	MacFaul", email: "kmacfaulj@oaic.gov.au" },
+  { id: 21, name: "Tanner	Janny", email: "tjannyk@cbc.ca" },
+  { id: 22, name: "Rudyard	Shakspeare", email: "rshakspearel@jiathis.com" },
+  { id: 23, name: "Vincents	Gaye", email: "vgayem@over-blog.com" },
+  { id: 24, name: "Findley	Witchell", email: "fwitchelln@ask.com" },
+  { id: 25, name: "Dell	McGlew", email: "dmcglewo@gov.uk" },
+  { id: 26, name: "Othella	Garrison", email: "ogarrisonp@stumbleupon.com" },
+  { id: 27, name: "Corissa	Duprey", email: "cdupreyq@i2i.jp" },
 ];
